Minor: Consolidate display related traits
